+++ axelor-quality/src/main/java/com/axelor/apps/quality/service/QualityControlService.java
@@ -22,7 +22,9 @@ import com.axelor.apps.quality.db.QualityControl;
 import com.axelor.apps.quality.db.QualityProcess;
 import com.axelor.exception.AxelorException;
 import com.google.inject.persist.Transactional;
+import java.io.IOException;
 import java.util.List;
+import javax.mail.MessagingException;
 
 public interface QualityControlService {
 
@@ -33,4 +35,8 @@ public interface QualityControlService {
   @Transactional
   public void preFillOperationsFromOptionals(
       QualityControl qualityControl, List<ControlPoint> optionalControlPointList);
+
+  void sendEmail(QualityControl qualityControl)
+      throws ClassNotFoundException, InstantiationException, IllegalAccessException,
+          MessagingException, IOException, AxelorException;
 }